python/requirements.txt
ask-sdk-core==1.17.1
ask-sdk-webservice-support==1.3.3
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
pytest==8.1.1
boto3==1.34.80
